+++ CHANGELOG.md
@@ -96,6 +96,8 @@ Ansible Changes By Release
   (https://github.com/ansible/ansible/pull/32342)
 * Ensure there is always a basdir so we always pickup group/host_vars
   https://github.com/ansible/ansible/pull/32269
+* Fix vars placement in ansible-inventory
+  https://github.com/ansible/ansible/pull/32276
 
 
 <a id="2.4.1"></a>